+++ src/commands/add.cmd.ts
@@ -1,4 +1,4 @@
-import JitsiBot from '../bot';
+import JitsiBot from '../jitsi-bot';
 import { IJimCommand } from '../command.service';
 import config from '../config';
 import { IIncomingMessage } from '../models/jitsi.interface';
@@ -15,7 +15,7 @@ export default <IJimCommand> {
         await jim.onAudioEnded();
       }
     }
-    jim.sendMessages(jim.queue.map((track) => track.title), event);
+    jim.sendMultilineMessage(jim.queue.map((track) => track.title), event);
   },
   description: '<url|searchTerm> - Add track to queue',
 };