+++ src/commands/help.cmd.ts
@@ -1,10 +1,10 @@
-import JitsiBot from '../bot';
+import JitsiBot from '../jitsi-bot';
 import { IJimCommand } from '../command.service';
 import { IIncomingMessage } from '../models/jitsi.interface';
 
 export default <IJimCommand> {
   execute: async (jim: JitsiBot, params: string[], event: IIncomingMessage) => {
-    await jim.sendMessages(
+    await jim.sendMultilineMessage(
       Object
         .keys(jim.cmdService.commands)
         .sort()