+++ src/commands/list.cmd.ts
@@ -1,4 +1,4 @@
-import JitsiBot from '../bot';
+import JitsiBot from '../jitsi-bot';
 import { IJimCommand } from '../command.service';
 import { IIncomingMessage } from '../models/jitsi.interface';
 
@@ -7,7 +7,7 @@ export default <IJimCommand> {
     if (jim.queue.length === 0) {
       await jim.sendMessage('No track in queue - Go ahead and add one :notes:', event);
     } else {
-      await jim.sendMessages(jim.queue.map((track) => track.title), event);
+      await jim.sendMultilineMessage(jim.queue.map((track) => track.title), event);
     }
   },
   description: 'Show tracks in queue',